+++ Motor2D/Tower.cpp
@@ -134,10 +134,10 @@ bool Tower::Update(float dt)
 			{
 				if (spell != nullptr)
 				{
-					DealDamage((entity->stats.power * spell->stats.damage_multiplicator) + ability->damage); // Spells control their own damage mutiplicator
+					DealDamage(((float)entity->stats.power * (float)spell->stats.damage_multiplicator) + (float)ability->damage); // Spells control their own damage mutiplicator
 				}
 				else
-					DealDamage((entity->stats.power * ability->damage_multiplicator) + ability->damage);
+					DealDamage(((float)entity->stats.power * (float)ability->damage_multiplicator) + (float)ability->damage);
 			}	
 
 			Die(entity);
